apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
supabase==2.9.1
openai==1.54.4
pypdf==5.1.0
tiktoken==0.8.0
python-multipart==0.0.12
ruff==0.7.4
